requirements.txt
dj-database-url==1.0.0
Django==4.1.2
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-geojson==4.0.0
django-leaflet==0.28.3
django-smart-selects==1.6.0
email-validator==1.3.0
Pillow==9.3.0
psycopg2==2.9.5
python-decouple==3.6